example/decorators.py
from rest_framework.response import Response
from rest_framework.views import status


def validate_request_data(fn):
    def decorated(*args, **kwargs):
        return fn(*args, **kwargs)
    return decorated
