requirements.txt
Django==2.0.3
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
PyJWT==1.6.1
pytz==2018.3
Pillow>=8.0.0,<9.0.0

# external requirements
# click
# Sphinx
# coverage
# awscli
# flake8
# python-dotenv>=0.5.1

##matplotlib==3.4.3
# sentinelsat==1.1.0
# googlemaps==4.5.3
# pandas==1.1.5
# seaborn>=0.11.0
# tabulate==0.8.9
# cython>=0.29.24
# opencv-python==4.5.4.58
# wheel>=0.37.0
# wandb>=0.12.6

# scipy==1.5.4
# tensorboard>=2.4.1
# albumentations>=1.1.0
##pycocotools==2.0.2
# pycocotools

# pytorch-lightning>=1.5.9
